% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPValue.R
\name{addPvalue}
\alias{addPvalue}
\title{addPvalue}
\usage{
addPvalue(
  pred,
  df = NULL,
  hypot = 0,
  estimate = "estimate",
  variance = "variance"
)
}
\arguments{
\item{variance}{}
}
\description{
addPvalue
}
