% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addUpperLower.R
\name{addUpperLower}
\alias{addUpperLower}
\title{addUpperLower}
\usage{
addUpperLower(
  pred,
  df = NULL,
  estimate = "estimate",
  variance = "variance",
  alpha = 0.05
)
}
\arguments{
\item{variance}{}
}
\description{
addUpperLower
}
