% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIC.R
\name{AIC.lnLik}
\alias{AIC.lnLik}
\title{AIC.lnLik}
\usage{
\method{AIC}{lnLik}(fit, k = 2)
}
\arguments{
\item{fit}{an lnLik structure}

\item{k}{the penalty parameter k, defaults to 2 influences only AIC}
}
\value{
a named vector containing the values for number of observations (nobs), number of paramters estimated (npar), AIC, AICc and BIC
}
\description{
AIC.lnLik
}
\examples{
AIC(fit) # where fit is an lnLik structure (result of lnLik())
}
