% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allDifferences.R
\name{allDifferences}
\alias{allDifferences}
\title{allDifferences}
\usage{
allDifferences(estimates, vcov, df = NULL, estimate = "estimate")
}
\arguments{
\item{estimates}{a data frame with one of the columns containing the estimates}

\item{vcov}{the covariance matrix of the estimates}

\item{df}{the residual degrees of freedom of the model}

\item{estimate}{the name of the estimate in the data.frame estimates}
}
\value{
a list with the matrix of pairwise differences, matrix with variance of differences (not the covariance) and the matrix of the P-values
}
\description{
allDifferences
}
\examples{
fit <- lm(y ~ agegroup*gender, data = data)
pred <- createPredictors(fit, marginal='yes')
est <- predictModel(fit, predictors=pred)
allDifferences(est$estimates, est$vcov, est$df)
}
