% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy.type.R
\name{copy.type}
\alias{copy.type}
\title{copy.type}
\usage{
copy.type(newdata, data, varnames)
}
\arguments{
\item{newdata}{a data matrix with columns varnames which has to match types of data\link{,varnames}}

\item{data}{a data matrix from which to copy the data type}

\item{varnames}{the names of the columns to check}
}
\value{
a data matrix with matching data types and levels for varnames with data\link{,varnames}
}
\description{
copy.type
}
\examples{
copy.type(estimates, data, c('AgeGroup','Ethnicity'))
}
