% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctable.R
\name{count}
\alias{count}
\title{count}
\usage{
count(x)
}
\arguments{
\item{x}{a vector with values}
}
\value{
the number of non-missing values in vector x
}
\description{
count
}
\examples{
count(c(1,2,3,4,NA,5))
}
