% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPredictors.R
\name{createPredictors}
\alias{createPredictors}
\title{createPredictors}
\usage{
createPredictors(
  object,
  data = NULL,
  marginal = c("yes", "no", "minmax", "existing"),
  nsteps = 1
)
}
\arguments{
\item{object}{a fitted model or a character string indicating the explanatory variables}

\item{data}{the dataset to create the predictors from, when omitted getData(object) is used)}

\item{marginal}{possible values c('yes','no','minmax','existing') see details section}

\item{nsteps}{integer value > 0, indicating the number of steps to use for creating a series of a continuous variable}
}
\value{
a new dataset with the predictor values as specified
}
\description{
createPredictors
}
\details{
Additional details...
}
\examples{
fit <- lm(x ~ Sex*AgeClass, data=data)
newData <- createPredictors(fit, marginal='yes')
}
