% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctable.R
\name{ctable}
\alias{ctable}
\title{ctable}
\usage{
ctable(
  data,
  variable,
  rows = NULL,
  cols = NULL,
  fun = "nObs",
  rdrop = TRUE,
  cdrop = TRUE,
  na.rm = TRUE,
  hsep = c(".", ":"),
  hkeep = FALSE,
  vpos = ifelse(length(variable) == 1, 0, 1),
  vname = "variable",
  fpos = ifelse(length(fun) == 1, ifelse(is.null(cols), Inf, 0), Inf),
  internal.sep = "\\b",
  ...
)
}
\arguments{
\item{data}{a data frame with data to make a table from}

\item{variable}{a (named) character vector with the variables to be summarised}

\item{rows}{a (named) character vector indicating the rows factors of the table}

\item{cols}{a (named) character vector indicating the column factors of the table}

\item{fun}{a (named) character vector with the function names}

\item{rdrop}{a logical indicting if the non existing row combinations should be dropped from the table}

\item{cdrop}{a logical indicting if the non existing column combinations should be dropped from the table}

\item{na.rm}{a logical indicating if the missing values should be ignored, the value is populated to the functions in fun if it has na.rm as argument}

\item{hsep}{a character string of two elements with the separators for the header, the first the separator between factorname and level, the second between the factors}

\item{hkeep}{a logical indicating to keep all factor names and levels in the header, only used in simple headers}

\item{vpos}{The position of the variables in the row classification}

\item{vname}{the name of the column for the variables}

\item{fpos}{the position of the functions in the column classification}

\item{internal.sep}{internal separation character used in the function interaction}

\item{...}{extra parameters to give to one or more of the functions in fun}
}
\value{
a data.frame of the class ctable with extra attributes for the header and row classifications
}
\description{
ctable
}
\examples{
meanAndSd <- function(x, na.rm=TRUE, digits=2, ...) {
  m <- mean(x, na.rm=na.rm)
  if (is.na(m)) {
    NA
  } else {
    paste0(round(mean(x, na.rm=na.rm), digits)," (",round(sd(x, na.rm=na.rm), digits), ")")
  }
}
#' ctable(data, 'length', rows='species', fun=c('Mean (sd)'='meanAndsd'), digits=3)

data(iris)
ctable(iris, c('Sepal.Length', 'Sepal.Width', 'Petal.Length', 'Petal.Width'), cols='Species', fun='mean')
}
