% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumcountby.R
\name{cumcountby}
\alias{cumcountby}
\title{cumcountby}
\usage{
cumcountby(data, newvar, var, fact.names)
}
\arguments{
\item{data}{a dataframe containing the classifying factors and variable name var}

\item{newvar}{The name of the newly created variable}

\item{var}{the variable to use as the ordering variable}

\item{fact.names}{the factor names of the calssification}
}
\value{
the data frame data with a new variable indicating the classification and a new variable indicating the cumulative counts
}
\description{
calculates the cumulative count within a classification over the values of a
variable and stores it in a new variable
}
\examples{
data <- cumcountby(data, 'cumcount', 'DurationInDays', c('Sex','Ethnicity'))
}
