% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractNames.R
\name{extractNames}
\alias{extractNames}
\title{extractNames}
\usage{
extractNames(x, split = "!\\"#$\%&'()*+,-/:;<=>?@[\\\\]^`{|}~ ")
}
\arguments{
\item{x}{input text to split}

\item{split}{punctuation characters, used for splitting the text recursively}
}
\value{
names and numbers contained in x, all punctuation (except _ and .) characters are removed
}
\description{
extractNames
}
\examples{
extractNames('Surv(Duration/365.25, Dead)')
}
