% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctable.R
\name{first}
\alias{first}
\title{first}
\usage{
first(x, na.rm = TRUE)
}
\arguments{
\item{x}{a vector with values}

\item{na.rm}{a logical indicating if missing values should be included}
}
\value{
the first element of the vector
}
\description{
first
}
\examples{
first(1:10)
}
