% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCoefficients.R
\name{getCoefficients.lnLik}
\alias{getCoefficients.lnLik}
\title{#' @export
getCoefficients.coxph <- function(fit) {
res <- as.data.frame(summary(fit)$coefficients)
names(res)\link{names(res)=='coef'} <- 'Estimate'
res
}}
\usage{
\method{getCoefficients}{lnLik}(fit)
}
\description{
#' @export
getCoefficients.coxph <- function(fit) {
res <- as.data.frame(summary(fit)$coefficients)
names(res)\link{names(res)=='coef'} <- 'Estimate'
res
}
}
