% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCoefficients.R
\name{getCoefficients}
\alias{getCoefficients}
\title{getCoefficients}
\usage{
getCoefficients(fit)
}
\arguments{
\item{fit}{a fitted model}
}
\value{
a dataframe with Estimates, Std. Error, a statistic and P values
}
\description{
getCoefficients
}
\examples{
fit <- lm(y ~ AgeGroup, data=data)
getCoefficients(fit)
}
