% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getData}
\alias{getData}
\title{getData}
\usage{
getData(fit)
}
\arguments{
\item{fit}{fitted model}
}
\value{
the original dataset used in the regression
}
\description{
getData
}
\examples{
fit <- lm(y ~ x, data=data)
getData(fit)
}
