% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDf.R
\name{getDf}
\alias{getDf}
\title{getDf}
\usage{
getDf(fit)
}
\arguments{
\item{fit}{a fitted model}
}
\value{
the residual degrees of freedom for that model
}
\description{
getDf
}
\examples{
fit <- lm(y ~ gender, data = data)
getDf(fit)
}
