% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDifferences.R
\name{getDifferences}
\alias{getDifferences}
\title{getDifferences}
\usage{
getDifferences(
  est,
  fact,
  fact.baseline = NULL,
  within = NULL,
  within.baseline = NULL,
  terms = c(fact, within),
  estimate = "estimate",
  sep = "."
)
}
\arguments{
\item{sep}{}
}
\description{
getDifferences
}
