% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFixedVars.R
\name{getFixedVars}
\alias{getFixedVars}
\title{getFixedVars}
\usage{
getFixedVars(fit, data = NULL)
}
\arguments{
\item{fit}{a fitted model}

\item{data}{data used, default NULL, extracted from the model with getData}
}
\value{
a character string with the variable names in the fixed model
}
\description{
getFixedVars
}
\examples{
fit <- lmer(count ~ gender*Age + (1|hospital), data=data)
getFixedVars(fit)
}
