% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHeader.R
\name{getHeader}
\alias{getHeader}
\title{getHeader}
\usage{
getHeader(colnames, sep1 = ";", sep2 = ":")
}
\arguments{
\item{colnames}{a crosstable or names(data.frame))}

\item{sep1}{symbol indicating the separation of variables}

\item{sep2}{symbol indicating the separation between variable and its level}
}
\value{
a data.frame with the header information
}
\description{
getHeader
}
