% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLink.R
\name{getLink}
\alias{getLink}
\title{getLink}
\usage{
getLink(fit)
}
\arguments{
\item{fit}{a fitted model}
}
\value{
the name of the link function used in the model
}
\description{
getLink
}
\examples{
fit <- glm(count ~ gender, family=poisson, data=data)
getLink(fit)
}
