% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLinkInv.R
\name{getLinkInv}
\alias{getLinkInv}
\title{getLinkInv}
\usage{
getLinkInv(fit)
}
\arguments{
\item{fit}{a fitted model}
}
\value{
the name of the link function used in the model
}
\description{
getLinkInv
}
\examples{
fit <- glm(count ~ gender, family=poisson, data=data)
getLinkInv(fit)
}
