% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getResponseVars.R
\name{getResponseVars}
\alias{getResponseVars}
\title{getResponseVars}
\usage{
getResponseVars(fit, data = NULL)
}
\arguments{
\item{fit}{a fitted model e.g. result of glm, coxph, etc.}

\item{data}{the data used in the model, otherwise it will be derived with getData()}
}
\value{
a character list of response variable names, the attr(,'full.description')
will point to the full formula for the left hand side, attr(,'full.list') will
list all names and numbers contained in attr(,'full.description')
}
\description{
getResponseVars
}
\examples{
fit <- coxph(Surv(duration, dead) ~ age, data=data)
getResponseVars(fit)
}
