% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSimulatedDifferences.R
\name{getSimulatedDifferences}
\alias{getSimulatedDifferences}
\title{getSimulatedDifferences}
\usage{
getSimulatedDifferences(
  est,
  fact,
  fact.baseline = NULL,
  within = NULL,
  within.baseline = NULL,
  terms = c(fact, within),
  estimate = "estimate",
  sep = ".",
  nsimulations = 10000
)
}
\arguments{
\item{nsimulations}{}
}
\description{
getSimulatedDifferences
}
