% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTerms.R
\name{getTerms}
\alias{getTerms}
\title{getTerms}
\usage{
getTerms(fit)
}
\arguments{
\item{fit}{a fitted model}
}
\value{
a terms object
}
\description{
getTerms
}
\examples{
fit <- lm(y ~ AgeGroup*Gender, data=data)
getTerms(fit)
}
