% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctable.R
\name{last}
\alias{last}
\title{first}
\usage{
last(x, na.rm = TRUE)
}
\arguments{
\item{x}{a vector with values}

\item{na.rm}{a logical indicating if missing values should be included}
}
\value{
the last element of the vector
}
\description{
first
}
\examples{
last(1:10)
}
