% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lnLik.R
\name{lnLik}
\alias{lnLik}
\title{lnLik}
\usage{
lnLik(
  fun,
  x,
  parinfo,
  data,
  trace = TRUE,
  control = lnLikControl(),
  sep = "_",
  n = NULL,
  ...
)
}
\arguments{
\item{fun}{The distribution function to be used in the log-likelihood function}

\item{x}{the name of the vector with the x variables (usually the quantiles)}

\item{parinfo}{the result of the Parinfo structure with the initial parameter values and relationships}

\item{data}{the data frame containing the x and possible n values and the variables used in the fact parameter of parinfo}

\item{trace}{a logical to give trace output at every step}

\item{control}{a lnLikControl structure}

\item{sep}{a separator value used in parinfo}

\item{n}{the name of the vector with the number of observations for each x value}

\item{...}{additional variable names to be used in fun}
}
\value{
an lnLik structure containing output
}
\description{
lnLik is used to fit a log likelihood function
}
