% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lnLikControl.R
\name{lnLikControl}
\alias{lnLikControl}
\title{Control of lnLik model fitting}
\usage{
lnLikControl(
  method = c("NewtonRaphson", "Levenberg", "Marquardt"),
  criterium = c("delta", "jacobian", "lnlik"),
  eps = 1e-07,
  nsteps = 200,
  hessian = c("calculate", "expected"),
  lambda = 1,
  lambda.multiplier = 10,
  reduce = TRUE,
  replaceNAN = FALSE,
  useGInv = FALSE
)
}
\arguments{
\item{method}{convergence method to use}

\item{criterium}{the stop criterium}

\item{eps}{the value when reached stops the iterations}

\item{nsteps}{the maximum number of steps in iteration}

\item{hessian}{which hessian to use, second order derived with Deriv, or expected one}

\item{lambda}{the lambda value in levenberg}

\item{lambda.multiplier}{the lambda multiplier in levenberg-marquardt}

\item{reduce}{A logical indicating to reduce the dataset if possible}

\item{replaceNAN}{replace NAN and Inf with zeros in derivatives}

\item{useGInv}{use generalised inverse}
}
\value{
The control function return a list containing
\enumerate{
\item general control parameters, method, hessian type
\item optimisation control: criterium, eps, nsteps
}
}
\description{
Construct control structures for lnLik model fitting. All arguments have defaults
}
\examples{
str(lnLikControl())
}
