% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lnLikInitModelParameters.R
\name{lnLikInitModelParameters}
\alias{lnLikInitModelParameters}
\title{lnLikInitModelParameters}
\usage{
lnLikInitModelParameters(p, models, parnames, weights = NULL)
}
\arguments{
\item{weights}{}
}
\description{
groups parameter values (mean or weighted mean) according to the (sub)models per parameter
}
\examples{
p <0 lnLikInitalParameters(data, 'Ethnicity.Sex.RHD', var='AgeInDays', initialWeibull)
fact <- list(k=~1+Ethnicity*Sex, lambda=~1+Ethnicity*RHD, c=~1)
par <- lnLikInitModelParameters(p, fact, attr(p, 'parnames', weights='nRecords'))

$k
(Intercept)         EthnicityITaukei           EthnicityOther                  SexMale EthnicityITaukei:SexMale   EthnicityOther:SexMale
4.540634332             -0.518679006             -0.564422963             -0.112425071              0.085505427              0.005916408

$lambda
(Intercept)      EthnicityITaukei        EthnicityOther                  RHD1 EthnicityITaukei:RHD1   EthnicityOther:RHD1
25264.71598            -334.40392              74.78567           -1173.46108           -5824.53872           -3892.44643

$c
(Intercept)
0
}
