% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lnLikPredictions.R
\name{lnLikPredictions}
\alias{lnLikPredictions}
\title{lnLikPredictions}
\usage{
lnLikPredictions(pred, fit, fun, nsimulations = 10000, vcov = FALSE)
}
\arguments{
\item{vcov}{}
}
\description{
lnLikPredictions
}
