% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctable.R
\name{meanAndSd}
\alias{meanAndSd}
\title{meanAndSd}
\usage{
meanAndSd(x, na.rm = TRUE, digits = 2, ...)
}
\arguments{
\item{x}{a vector with values}

\item{na.rm}{a logical indicating if missing values should be included}

\item{digits}{the number of digits to round the values before making a text string}

\item{...}{additional parameters, needed for use in function ctable}
}
\value{
a text with 'mean (sd)' with the number of digits specified
}
\description{
meanAndSd
}
\examples{
meanAndSd(c(1,2,3,4,NA,5), digits=4)
}
