% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull.R
\name{pdfWeibull}
\alias{pdfWeibull}
\title{The Weibull Distribution}
\usage{
pdfWeibull(x, k, lambda, c)
}
\arguments{
\item{x}{vector of quantiles}

\item{k}{shape paramter}

\item{lambda}{scale paramter}

\item{c}{offset parameter}
}
\description{
Density (pdf) and cumulative density (cdf) Weibull distribution functions with
parameters shape (k), scale(lambda) and location (c). The functions are well suited
to be used in the lnLik function as their derivatives have already been derived
pdfWeibull, cdfWeibull are in principle reparamaterisations of the dweibull and
qweibull of the stats package. In addition moment estimaters can be made using the
function initWeibull. Summary statistics can be derived from the parameters using
meanWeibull, medianWeibull, modeWeibull. The hazard function hazardWeibull has
been added too.
}
\details{
pdfWeibull(x, k, lambda, c)
cdfWeibull(x, k, lambda, c)
initialWeibull(x)
meanWeibull(k, lambda, c)
medianWeibull(k, lambda, c)
modeWeibull(k, lambda, c)
hazardWeibull(x, k, lamba, c)#'
}
\examples{
pdf(0:120, 4, 60, 0)

}
