% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictModel.R
\name{predictModel}
\alias{predictModel}
\title{predictModel}
\usage{
predictModel(fit, predictors, weights = NULL, df = NULL)
}
\arguments{
\item{fit}{a fitted model}

\item{predictors}{newData for which predictions and variances are needed}

\item{weights}{a numeric vector with weights for each record in the predictors}

\item{df}{the number of residual degrees of freedom, used for calculating the 95\% confidence limits. If omitted the function getDf(fit) will be used.}
}
\value{
a list with predictions, covariance matrix of the predictions, family, link and df
}
\description{
predictModel
}
\examples{
fit <- lm( y ~ ageclass*gender, data=data)
pred <- createPredictors(fit)
estimates <- predictmodel(fit, predictors=pred)
}
