% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processFormula.R
\name{processFormula}
\alias{processFormula}
\title{processFormula}
\usage{
processFormula(formula)
}
\arguments{
\item{formula}{a formula or a text representing a formula}
}
\value{
left and right hand site variables and terms and random variables or terms
}
\description{
processFormula
}
\examples{
processFormula('cbind(n, population) ~ Sex + log(Age) + (1|hospital)')
}
