% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replaceValues.R
\name{replacevalues}
\alias{replacevalues}
\title{replacevalues}
\usage{
replacevalues(vec, oldvalues, newvalues, warnings = TRUE, droplevels = TRUE)
}
\arguments{
\item{vec}{vector with values, can be a factor}

\item{oldvalues}{oldvalues to change, there shouldn't be double values in oldvalues}

\item{newvalues}{newvalues to replace matching oldvalues, length of newvalues must be equal to length of oldvalues or 1}

\item{warnings}{if a value in oldvalue is not in vec, a warning is given}

\item{droplevels}{if after replacement values are not in a factor then those levels will be dropped according to the value of droplevels (TRUE, FALSE)}
}
\value{
a vector or factor with new values, the levels of the factor will be expanded with new values (if the corresponding old value is in original ved) and old values will be dropped
}
\description{
replaces values in a vector/factor with new values. If the inputvector is a factor thant the levels of the factor will be adjusted.
Replacevalues can be used to swap values e.g. changing the order in likert scale, values are not recycled!
}
\examples{

replacevalues(vec=factor(c('x',NaN,NA,NA,'y')), oldvalues=c(NaN, NA), newvalues=c('Q')) # missing level Q will be added
replacevalues(vec=factor(c('x',NaN,NA,NA,'y')), oldvalues=c('d'), newvalues=c('Q'))  # level Q will not be added

}
