% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setZero.R
\name{setZero}
\alias{setZero}
\title{setZero}
\usage{
setZero(x)
}
\arguments{
\item{x}{a data structure in r}
}
\value{
any x value is changed if abs(x_value) <= TOL
}
\description{
setZero
}
\examples{
x <- setZero(x)
}
