% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitMatrix.R
\name{splitMatrix}
\alias{splitMatrix}
\alias{is.splitMatrix}
\alias{as.splitMatrix}
\title{splitMatrices}
\usage{
splitMatrix(mat, rows = 2, cols = 2, transpose = FALSE)

is.splitMatrix(mat)

as.splitMatrix(mat, ...)
}
\arguments{
\item{mat}{A matrix to split or a splitMatrix to re-split}

\item{rows}{A factor with length equal to the number of rows of the input matrix. If a single number is provided, it indicates the number of rows the matrix should be split.}

\item{cols}{A factor with length equal to the number of columns of the input matrix. If a single number is provided, it indicates the number of columns the matrix should be split.}

\item{transpose}{creates the sub matrices transpose}

\item{...}{Additional arguments passed to the function controlling the conversion process.}
}
\value{
a splitMatrix containing a matrix of matrices
}
\description{
Splits a matrix or splitMatrix into a matrix of matrices along rows or columns based on a number or factor.
}
\examples{
mat <- matrix(runif(100), 10, 10)
splitted.mat <- splitMatrix(mat, 2,2)
}
