% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_values.R
\name{missing_values}
\alias{missing_values}
\title{Descriptive and Visual Missing Value Assesment}
\usage{
missing_values(df, color = "#79E1BE", all = FALSE)
}
\arguments{
\item{color}{Character string indicating the color for missing values. Default: "#79E1BE"}

\item{all}{Logical parameter that shows all variables including those without missing values. Default: FALSE.}

\item{data}{Dataframe containing the variables to be analyzed.}
}
\value{
Prints results to console and invisibly returns a list with descriptive statistics and ggplot objects.
}
\description{
Provides descriptive statistics and visualizations of missing values in a dataframe.
}
\examples{
# Clinical dataset with missing values
clinical_df <- clinical_data(dropout = 0.1, missing = 0.05)

# Missing value analysis of only variables with missing values
missing_values(clinical_df)

# Show all variables including those without missing values
missing_values(clinical_df, all = TRUE)

}
