% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omnibus.R
\name{omnibus}
\alias{omnibus}
\title{Tests for comparing three or more groups (omnibus)}
\usage{
omnibus(
  dependent_var = NULL,
  independent_var = NULL,
  data = NULL,
  paired_var = NULL,
  alpha = 0.05,
  method = c("holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none"),
  na.action = c("na.omit", "na.exclude", "na.pass", "na.fail")
)
}
\arguments{
\item{dependent_var}{Character string. Name of the column that contains the data of the dependent variable.}

\item{independent_var}{Character string. Name of the column that contains the data of the independent variable.}

\item{data}{Character string. Name of the dataframe where the independent, dependent and paired variables are found.}

\item{paired_var}{Character string. Name of the column that contains the data of the paired variable. If paired,
the dataframe must have each group organized in these two options: (A, B, C, A, B, C, A, B, C) or (A, A, A, B, B, B, C, C, C)}

\item{alpha}{Numeric. Type I error rate (significance level). Default is 0.05.}

\item{method}{Character string. Method wanted to conduct a multiplicity correction.}

\item{na.action}{What to do when NAs are found in data.}
}
\description{
Omnibus tests evaluate overall effects in the given data. One-way ANOVA,
Repeated measures ANOVA, Kruskal Wallis test and Friedman test are included
in this term.
}
\examples{
# Examples go here
}
